+++ test/unit/rotate.js
@@ -35,6 +35,23 @@ describe('Rotation', function() {
     });
   });
 
+  it('Rotate by 270 degrees, square output ignoring aspect ratio', function(done) {
+    sharp(fixtures.inputJpg)
+      .resize(240, 240)
+      .ignoreAspectRatio()
+      .rotate(270)
+      .toBuffer(function(err, data, info) {
+        if (err) throw err;
+        assert.strictEqual(240, info.width);
+        assert.strictEqual(240, info.height);
+        sharp(data).metadata(function(err, metadata) {
+          assert.strictEqual(240, metadata.width);
+          assert.strictEqual(240, metadata.height);
+          done();
+        });
+    });
+  });
+
   it('Input image has Orientation EXIF tag but do not rotate output', function(done) {
     sharp(fixtures.inputJpgWithExif)
       .resize(320)